Refactor generator into helper classes
